Add the image of the sign in the list view.
This adds a field.image handler for the list view, so when specifying
an image field, the actual image shows (30px square for now) in the
tree and list views.

The handling of this is twofold:
 1. specify widget="image" in the view definition
 2. the javascript in static/src/js/view_list.js adds the code for it.

The code for this should probably be in OpenERP core, or at least in a
separate, albeit tiny, module.